[ODPR] - WIP POC Updates REG-2 Screen Flow
